# API-Engine-distro/API-Engine-distro-data/src/main/resources/local/swagger.yaml
swagger: '2.0'
info:
  version: v1
  title: API Engine
basePath: /API-Engine-web/v1
...
schemes:
  - http
paths:
...
  /gateways:
    get:
      tags:
        - gateways
      summary: List All Gateways
      description: This endpoint returns a list of all the Gateways that have been configured.
      operationId: list
      produces:
        - application/json
      parameters: []
      responses:
        '200':
          description: A list of configured Gateways.
          schema:
            type: array
            items:
              $ref: '#/definitions/GatewaySummaryBean'
    post:
      tags:
        - gateways
      summary: Create a Gateway
      description: This endpoint is called to create a new Gateway.
      operationId: create
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/NewGatewayBean'
      responses:
        '200':
          description: The newly created Gateway.
          schema:
            $ref: '#/definitions/GatewayBean'
    put:
      tags:
        - gateways
      summary: Test a Gateway
      description: "This endpoint is used to test the Gateway's settings prior to either creating or updating it.  The information will be used to attempt to create a link between the API Manager and the Gateway, by simply trying to ping the Gateway's status endpoint."
      operationId: test
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/NewGatewayBean'
      responses:
        '200':
          description: The result of testing the Gateway settings.
          schema:
            $ref: '#/definitions/GatewayTestResultBean'
  '/gateways/{gatewayId}':
    get:
      tags:
        - gateways
      summary: Get a Gateway by ID
      description: Call this endpoint to get the details of a single configured Gateway.
      operationId: get
      produces:
        - application/json
      parameters:
        - name: gatewayId
          in: path
          required: true
          type: string
      responses:
        '200':
          description: 'The Gateway identified by {gatewayId}.'
          schema:
            $ref: '#/definitions/GatewayBean'
    put:
      tags:
        - gateways
      summary: Update a Gateway
      description: "Use this endpoint to update an existing Gateway.  Note that the name of the Gateway cannot be changed, as the name is tied closely with the Gateway's ID.  If you wish to rename the Gateway you must delete it and create a new one."
      operationId: update
      consumes:
        - application/json
      parameters:
        - name: gatewayId
          in: path
          required: true
          type: string
        - in: body
          name: body
          required: false
          schema:
            $ref: '#/definitions/UpdateGatewayBean'
      responses:
        '204':
          description: 'successful, no content'
    delete:
      tags:
        - gateways
      summary: Delete a Gateway
      description: This endpoint deletes a Gateway by its unique ID.
      operationId: delete
      parameters:
        - name: gatewayId
          in: path
          required: true
          type: string
      responses:
        '204':
          description: 'successful, no content'
...
definitions:
...
  NewGatewayBean:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      type:
        type: string
        enum:
          - REST
      configuration:
        type: string
...
  GatewayBean:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      createdBy:
        type: string
      createdOn:
        type: string
        format: date-time
      modifiedBy:
        type: string
      modifiedOn:
        type: string
        format: date-time
      type:
        type: string
        enum:
          - REST
      configuration:
        type: string
...
  UpdateGatewayBean:
    type: object
    properties:
      description:
        type: string
      type:
        type: string
        enum:
          - REST
      configuration:
        type: string
...
  GatewaySummaryBean:
    type: object
    properties:
      id:
        type: string
      name:
        type: string
      description:
        type: string
      type:
        type: string
        enum:
          - REST
...
  GatewayTestResultBean:
    type: object
    properties:
      success:
        type: boolean
        default: false
      detail:
        type: string
...